requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.9.30=h06a4308_1
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
chardet=4.0.0=py37h06a4308_1003
conda=4.10.3=py37h06a4308_0
conda-build=3.21.5=py37h06a4308_0
conda-package-handling=1.7.3=py37h27cfd23_1
cryptography=3.4.8=py37hd23ed53_0
cudatoolkit=11.3.1=ha36c431_9
decorator=5.1.0=pyhd3eb1b0_0
dnspython=2.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pyhd3eb1b0_1
freetype=2.10.4=h5ab3b9f_0
giflib=5.2.1=h7b6447c_0
glob2=0.7=pyhd3eb1b0_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipython=7.27.0=py37hb070fc8_0
jedi=0.18.0=py37h06a4308_1
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libarchive=3.4.2=h62408e4_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblief=0.10.1=he6710b0_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py37h27cfd23_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
parso=0.8.2=pyhd3eb1b0_0
patchelf=0.13=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37h5aabda8_0
pip=21.0.1=py37h06a4308_0
pkginfo=1.7.1=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.8.0=py37h27cfd23_1
ptyprocess=0.7.0=pyhd3eb1b0_2
py-lief=0.10.1=py37h403a769_0
pycosat=0.6.3=py37h27cfd23_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pysocks=1.7.1=py37_1
python=3.7.11=h12debd9_0
python-etcd=0.4.5=pypi_0
python-libarchive-c=2.9=pyhd3eb1b0_1
pytorch=1.10.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=py37h27cfd23_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
ripgrep=12.1.1=0
ruamel_yaml=0.15.100=py37h27cfd23_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
soupsieve=2.2.1=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tk=8.6.10=hbc83047_0
torchelastic=0.2.0=pypi_0
torchtext=0.11.0=py37
torchvision=0.11.0=py37_cu113
tqdm=4.61.2=pyhd3eb1b0_1
traitlets=5.1.0=pyhd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pyhd3eb1b0_1
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
